backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
firebase-admin==6.4.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
statsmodels==0.14.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
torch==2.5.1
torchvision==0.20.1
Pillow
email-validator
ultralytics>=8.0.0
opencv-python>=4.8.0
